Refactor models files to be in separate package
