+++ library/src/main/java/com/wultra/android/digitalonboarding/VerificationService.kt
@@ -21,21 +21,21 @@ package com.wultra.android.digitalonboarding
 import android.content.Context
 import com.wultra.android.digitalonboarding.VerificationStatusNextStep.Value
 import com.wultra.android.digitalonboarding.log.WDOLogger
-import com.wultra.android.digitalonboarding.networking.ConsentApproveResponse
-import com.wultra.android.digitalonboarding.networking.ConsentResponse
+import com.wultra.android.digitalonboarding.networking.model.ConsentApproveResponse
+import com.wultra.android.digitalonboarding.networking.model.ConsentResponse
 import com.wultra.android.digitalonboarding.networking.CustomerOnboardingApi
 import com.wultra.android.digitalonboarding.networking.CustomerVerificationApi
-import com.wultra.android.digitalonboarding.networking.Document
-import com.wultra.android.digitalonboarding.networking.DocumentStatus
-import com.wultra.android.digitalonboarding.networking.DocumentSubmitResponse
-import com.wultra.android.digitalonboarding.networking.DocumentsStatusResponse
-import com.wultra.android.digitalonboarding.networking.IdentityVerificationStatus
-import com.wultra.android.digitalonboarding.networking.OTPDetailResponse
-import com.wultra.android.digitalonboarding.networking.PresenceCheckResponse
-import com.wultra.android.digitalonboarding.networking.ResendOtpResponse
-import com.wultra.android.digitalonboarding.networking.SDKInitResponse
-import com.wultra.android.digitalonboarding.networking.VerificationStatusResponse
-import com.wultra.android.digitalonboarding.networking.VerifyOtpResponse
+import com.wultra.android.digitalonboarding.networking.model.Document
+import com.wultra.android.digitalonboarding.networking.model.DocumentStatus
+import com.wultra.android.digitalonboarding.networking.model.DocumentSubmitResponse
+import com.wultra.android.digitalonboarding.networking.model.DocumentsStatusResponse
+import com.wultra.android.digitalonboarding.networking.model.IdentityVerificationStatus
+import com.wultra.android.digitalonboarding.networking.model.OTPDetailResponse
+import com.wultra.android.digitalonboarding.networking.model.PresenceCheckResponse
+import com.wultra.android.digitalonboarding.networking.model.ResendOtpResponse
+import com.wultra.android.digitalonboarding.networking.model.SDKInitResponse
+import com.wultra.android.digitalonboarding.networking.model.VerificationStatusResponse
+import com.wultra.android.digitalonboarding.networking.model.VerifyOtpResponse
 import com.wultra.android.powerauth.networking.IApiCallResponseListener
 import com.wultra.android.powerauth.networking.data.StatusResponse
 import com.wultra.android.powerauth.networking.error.ApiError
