+++ library/src/main/java/com/wultra/android/digitalonboarding/networking/CustomerVerificationApi.kt
@@ -18,6 +18,32 @@ package com.wultra.android.digitalonboarding.networking
 
 import android.content.Context
 import com.wultra.android.digitalonboarding.Utils
+import com.wultra.android.digitalonboarding.networking.model.CancelRequest
+import com.wultra.android.digitalonboarding.networking.model.ConsentApproveRequest
+import com.wultra.android.digitalonboarding.networking.model.ConsentApproveResponse
+import com.wultra.android.digitalonboarding.networking.model.ConsentRequest
+import com.wultra.android.digitalonboarding.networking.model.ConsentResponse
+import com.wultra.android.digitalonboarding.networking.model.DocumentSubmitRequest
+import com.wultra.android.digitalonboarding.networking.model.DocumentSubmitRequestData
+import com.wultra.android.digitalonboarding.networking.model.DocumentSubmitResponse
+import com.wultra.android.digitalonboarding.networking.model.DocumentsStatusRequest
+import com.wultra.android.digitalonboarding.networking.model.DocumentsStatusResponse
+import com.wultra.android.digitalonboarding.networking.model.EmptyRequest
+import com.wultra.android.digitalonboarding.networking.model.OTPDetailRequest
+import com.wultra.android.digitalonboarding.networking.model.OTPDetailRequestData
+import com.wultra.android.digitalonboarding.networking.model.OTPDetailResponse
+import com.wultra.android.digitalonboarding.networking.model.OTPDetailType
+import com.wultra.android.digitalonboarding.networking.model.PresenceCheckRequest
+import com.wultra.android.digitalonboarding.networking.model.PresenceCheckResponse
+import com.wultra.android.digitalonboarding.networking.model.PresenceCheckSubmitRequest
+import com.wultra.android.digitalonboarding.networking.model.ResendOtpResponse
+import com.wultra.android.digitalonboarding.networking.model.SDKInitRequest
+import com.wultra.android.digitalonboarding.networking.model.SDKInitResponse
+import com.wultra.android.digitalonboarding.networking.model.StartRequest
+import com.wultra.android.digitalonboarding.networking.model.VerificationResendOtpRequest
+import com.wultra.android.digitalonboarding.networking.model.VerificationStatusResponse
+import com.wultra.android.digitalonboarding.networking.model.VerifyOtpRequest
+import com.wultra.android.digitalonboarding.networking.model.VerifyOtpResponse
 import com.wultra.android.powerauth.networking.Api
 import com.wultra.android.powerauth.networking.EndpointBasic
 import com.wultra.android.powerauth.networking.EndpointSigned
